Trait kilt_support::signature::VerifySignature

pub trait VerifySignature {
    type SignerId;
    type Payload;
    type Signature;

    // Required methods
    fn verify(
        signer: &Self::SignerId,
        payload: &Self::Payload,
        signature: &Self::Signature
    ) -> SignatureVerificationResult;
    fn weight(payload_byte_length: usize) -> Weight;
}

A signature verification implementation.

Required Associated Types

type SignerId

The identifier of the signer.

type Payload

The type of the payload that can be verified with the implementation.

type Signature

The type of the signature that is expected by the implementation.

Required Methods

fn verify( signer: &Self::SignerId, payload: &Self::Payload, signature: &Self::Signature ) -> SignatureVerificationResult

Verifies that the signature matches the payload and has been generated by the signer.

fn weight(payload_byte_length: usize) -> Weight

The weight if the signature verification.