Trait kilt_support::traits::CallSources

pub trait CallSources<S, P> {
    // Required methods
    fn sender(&self) -> S;
    fn subject(&self) -> P;
}

The sources of a call struct.

This trait allows to differentiate between the sender of a call and the subject of the call. The sender account submitted the call to the chain and might pay all fees and deposits that are required by the call.

Required Methods

fn sender(&self) -> S

The sender of the call who will pay for all deposits and fees.

fn subject(&self) -> P

The subject of the call.